+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 Changelog
 =========
 
+v1.0.1 (2022-01-18)
+-------------------
+
+- Better code with a structure Entry
+
 v1.0.0 (2022-01-17)
------------------
+-------------------
 
 Users can:
 - Choose a master password
 - Add a new entry : a couple of secret information
-- Print actual entries
+- Print current entries
 - Search and print entries according a pattern